pyproject.toml
[project]
name = "pd-discovery-platform"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncio>=3.4.3",
    "crewai>=0.134.0",
    "crewai-tools>=0.48.0",
    "fastmcp>=2.9.2",
    "firecrawl>=2.12.0",
    "google-generativeai>=0.8.5",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.3",
    "numpy>=2.3.1",
    "omegaconf>=2.3.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.0",
    "pydantic>=2.11.7",
    "pytest-asyncio>=1.0.0",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "scikit-learn>=1.7.0",
    "scipy>=1.16.0",
    "seaborn>=0.13.2",
    "statsmodels>=0.14.4",
]
